backend/requirements.txt
absl-py==2.3.1
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
bidict==0.23.1
blinker==1.9.0
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
contourpy==1.3.3
cycler==0.12.1
distro==1.9.0
dnspython==2.7.0
eventlet==0.40.3
fastapi==0.116.1
flask==3.1.2
flask-socketio==5.5.1
flatbuffers==25.2.10
fonttools==4.59.2
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.179.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
greenlet==3.2.4
grpcio==1.74.0
grpcio-status==1.62.3
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
jax==0.7.1
jaxlib==0.7.1
jinja2==3.1.6
jiter==0.10.0
kiwisolver==1.4.9
loguru==0.7.3
markupsafe==3.0.2
matplotlib==3.10.5
mediapipe==0.10.14
ml-dtypes==0.5.3
numpy==2.2.6
openai==1.102.0
opencv-contrib-python==4.12.0.88
opencv-python==4.12.0.88
opt-einsum==3.4.0
packaging==25.0
pillow==11.3.0
proto-plus==1.26.1
protobuf==4.25.8
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic-core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-engineio==4.12.2
python-multipart==0.0.20
python-socketio==5.13.0
pyyaml==6.0.2
rapidfuzz==3.14.0
requests==2.32.5
rsa==4.9.1
scipy==1.16.1
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.2
starlette==0.47.3
tqdm==4.67.1
typing-extensions==4.15.0
typing-inspection==0.4.1
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
werkzeug==3.1.3
wsproto==1.2.0
